requirements.txt
fastapi[standard]
uvicorn[standard]
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic
python-dotenv
strawberry-graphql
openai
gql
requests_toolbelt
aiohttp